+++ src/chart/generator/axis.cpp
@@ -93,8 +93,9 @@ bool DimensionAxis::operator==(const DimensionAxis &other) const
 	return enabled == other.enabled && values == other.values;
 }
 
-void DimensionAxis::setLabels(double step)
+bool DimensionAxis::setLabels(double step)
 {
+	bool hasLabel{};
 	step = std::max(step, 1.0);
 	double currStep = 0.0;
 
@@ -104,7 +105,9 @@ void DimensionAxis::setLabels(double step)
 		if (++curr <= currStep) continue;
 		currStep += step;
 		item.label = item.categoryValue;
+		hasLabel = true;
 	}
+	return hasLabel;
 }
 
 DimensionAxis interpolate(const DimensionAxis &op0,