+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - next() can be called multiple times from Plugin hooks
 - Line and circle chats with only dimensions on x, and y axes the markers were off the axis labels.
 - Crash on TreeMap only with negative values
+- On dimension axis where no marker, print the dimension name as default title.
+- On measure axis with specified channel min-max where no marker, print the labels.
 
 ### Added